+++ alphadia/plexscoring/features/fragment_features.py
@@ -3,7 +3,7 @@ import numpy as np
 
 from alphadia import utils
 from alphadia.numba.numeric import fragment_correlation, fragment_correlation_different
-from alphadia.plexscoring.features_.features_utils import (
+from alphadia.plexscoring.features.features_utils import (
     cosine_similarity_a1,
     weighted_center_mean_2d,
 )
