+++ alphadia/plexscoring/features/reference_features.py
@@ -1,7 +1,7 @@
 import numba as nb
 import numpy as np
 
-from alphadia.plexscoring.features_.features_utils import (
+from alphadia.plexscoring.features.features_utils import (
     cosine_similarity_a1,
 )
 from alphadia.utils import USE_NUMBA_CACHING
